+++ src/main/java/com/zerobase/cafebom/pay/service/dto/OrdersAddDto.java
@@ -1,7 +1,7 @@
 package com.zerobase.cafebom.pay.service.dto;
 
-import com.zerobase.cafebom.pay.controller.form.AddOrdersForm;
-import com.zerobase.cafebom.pay.controller.form.AddOrdersForm.OrderedProductForm;
+import com.zerobase.cafebom.pay.controller.form.OrdersAddForm;
+import com.zerobase.cafebom.pay.controller.form.OrdersAddForm.OrderedProductForm;
 import java.util.List;
 import java.util.stream.Collectors;
 import lombok.Builder;
@@ -9,14 +9,14 @@ import lombok.Getter;
 
 @Getter
 @Builder
-public class AddOrdersDto {
+public class OrdersAddDto {
 
     private String payment; // enum 인데 임시로 스트링
 
     private List<OrderedProductDto> products;
 
-    public static AddOrdersDto from(AddOrdersForm form) {
-        return AddOrdersDto.builder()
+    public static OrdersAddDto from(OrdersAddForm form) {
+        return OrdersAddDto.builder()
             .payment(form.getPayment())
             .products(
                 form.getProducts()
